templates/azure-bot/src/api/requirements.txt
python-dotenv
fastapi>=0.115.11
uvicorn>=0.34.0
teams-ai>=1.4.0,<2.0.0
semantic-kernel==1.22.0
